test/inspector/runtime/call-function-on-async-expected.txt

Tests that Runtime.callFunctionOn works with awaitPromise flag.

Running test: testArguments
{
    result : {
        type : string
        value : undefined|NaN|[object Object]|[object Object]
    }
}

Running test: testSyntaxErrorInFunction
{
    exceptionDetails : {
        columnNumber : 2
        exception : {
            className : SyntaxError
            description : SyntaxError: Unexpected token }
            objectId : 0
            subtype : error
            type : object
        }
        exceptionId : 0
        lineNumber : 1
        scriptId : 0
        text : Uncaught
    }
    result : {
        className : SyntaxError
        description : SyntaxError: Unexpected token }
        objectId : [ObjectId]
        subtype : error
        type : object
    }
}

Running test: testExceptionInFunctionExpression
{
    exceptionDetails : {
        columnNumber : 15
        exception : {
            className : Error
            description : Error     at <anonymous>:1:22     at <anonymous>:1:36
            objectId : 0
            subtype : error
            type : object
        }
        exceptionId : 0
        lineNumber : 0
        scriptId : 0
        text : Uncaught
    }
    result : {
        className : Error
        description : Error     at <anonymous>:1:22     at <anonymous>:1:36
        objectId : [ObjectId]
        subtype : error
        type : object
    }
}

Running test: testFunctionReturnNotPromise
{
    code : -32000
    message : Result of the function call is not a promise
}

Running test: testFunctionReturnResolvedPromiseReturnByValue
{
    result : {
        type : object
        value : {
            a : 3
        }
    }
}

Running test: testFunctionReturnResolvedPromiseWithPreview
{
    result : {
        className : Object
        description : Object
        objectId : [ObjectId]
        preview : {
            description : Object
            overflow : false
            properties : [
                [0] : {
                    name : a
                    type : number
                    value : 3
                }
            ]
            type : object
        }
        type : object
    }
}

Running test: testFunctionReturnRejectedPromise
{
    exceptionDetails : {
        columnNumber : 0
        exception : {
            objectId : 0
            type : object
            value : {
                a : 3
            }
        }
        exceptionId : 0
        lineNumber : 0
        stackTrace : {
            callFrames : [
            ]
        }
        text : Uncaught (in promise)
    }
    result : {
        type : object
        value : {
            a : 3
        }
    }
}
